# autoevolve-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Company Cockpit",
    "version": "0.1.0"
  },
  "paths": {
...
    "/company/setup/providers": {
      "get": {
        "tags": [
          "setup"
        ],
        "summary": "Setup Providers",
        "operationId": "setup_providers_company_setup_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/company/setup/keys/test": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Test Setup Keys",
        "operationId": "test_setup_keys_company_setup_keys_test_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupKeysPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupKeyTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/company/setup/keys": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Save Setup Keys",
        "operationId": "save_setup_keys_company_setup_keys_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupKeysPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupKeySaveResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/company/setup/step": {
      "get": {
        "tags": [
          "setup"
        ],
        "summary": "Get Setup Step Route",
        "operationId": "get_setup_step_route_company_setup_step_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Set Setup Step Route",
        "operationId": "set_setup_step_route_company_setup_step_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupStepPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SetupKeySaveResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "saved": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Saved"
          },
          "backup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backup"
          },
          "restart_required": {
            "type": "boolean",
            "title": "Restart Required"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "group",
          "saved",
          "backup",
          "restart_required"
        ],
        "title": "SetupKeySaveResult"
      },
      "SetupKeyTestResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "checked": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Checked"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "checked"
        ],
        "title": "SetupKeyTestResult"
      },
      "SetupKeysPayload": {
        "properties": {
          "group": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Group"
          },
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "group"
        ],
        "title": "SetupKeysPayload"
      },
      "SetupStepPayload": {
        "properties": {
          "step": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Step"
          }
        },
        "type": "object",
        "required": [
          "step"
        ],
        "title": "SetupStepPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}